feat: implement singleton mixin for NodeDedupManager and SerializableFactory; refactor serialization methods and update scene loading logic

+++ kirara-dance/Scene/Animation.h
@@ -34,7 +34,7 @@ template <typename T> struct AnimationKey {
     void serialize(auto &ar) {
         ar(cereal::make_nvp("time", time));
         ar(cereal::make_nvp("value", value));
-        ar(cereal::make_nvp("interp", static_cast<uint8_t>(interp)));
+        ar(cereal::make_nvp("interp", interp));
     }
 };
